Revert "Kanji test"
