--- .github/pull_request_template.md
@@ -1,38 +1,0 @@
-## 🔍 이 PR로 해결하고자 하는 문제는 무엇인가요?
-
-> 구현한 기능이나 해결한 과제, PR을 열게 된 계기나 목적을 리뷰어가 이해하기 쉽게 작성해주세요.
-> 백로그 링크, 다이어그램, Figma 등의 관련 문서를 첨부하면 더 좋습니다.
-
-## ✨ 이 PR로 주요하게 바뀐 점은 무엇인가요?
-
-> 문제를 해결하면서 주요 변경 사항을 작성해주세요.
-
--
-
-## 🔖 주요 변경 사항 외에 추가로 변경된 부분이 있나요?
-
-> 없으면 "없음"이라고 작성해주세요.
-
--
-
-## 🙏🏻 리뷰어가 특히 봐주었으면 하는 부분은 무엇인가요?
-
-> 개발 과정에서 다른 사람의 의견이 궁금하거나, 크로스 체크가 필요하다고 느낀 코드가 있으면 알려주세요.
-
--
-
-## 🩺 이 PR로 테스트나 검증이 필요한 부분이 있나요?
-
-> 테스트가 필요한 항목이나 테스트 코드가 추가되었다면 함께 작성해주세요.
-
--
-
-## 📚 관련된 Issue나 Notion, 문서
-
-> 이 PR이 해결하려는 문제와 관련된 Issue나 문서, Notion이 있다면 링크를 작성해주세요. Notion의 경우 [제목](링크) 형식으로 첨부해주세요.
-
--
-
-## 🖥 작동하는 모습
-
-> 스크린샷이나 녹화된 비디오, 또는 gif를 추가해서, 리뷰어가 변경점을 이해하는 데 도움이 되도록 해주세요.
